# ollama-distributed/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OllamaMax Distributed API
  description: |
    OllamaMax Distributed is an enterprise-grade distributed AI platform that transforms 
    the single-node Ollama architecture into a horizontally scalable, fault-tolerant system.
    
    ## Authentication
    
    The API uses JWT-based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <JWT_TOKEN>
    ```
    
    ## Base URLs
    
    - **API Server**: `http://localhost:8080`
    - **Web Interface**: `http://localhost:8081`
    - **Metrics**: `http://localhost:9090`
    
    ## Rate Limiting
    
    Default rate limit: 1000 requests per minute per client.
    
  version: 1.0.0
  contact:
    name: OllamaMax Support
    url: https://github.com/khryptorgraphics/ollamamax
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.ollamamax.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/models:
    get:
      summary: List available models
      description: Get a list of all available AI models in the cluster
      tags:
        - Models
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of models per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
...
  /api/v1/generate:
    post:
      summary: Text generation
      description: Generate text using a specified model
      tags:
        - Inference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
            example:
              model: "llama2:7b"
              prompt: "Explain quantum computing"
              max_tokens: 500
              temperature: 0.7
      responses:
        '200':
          description: Generated text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ModelsResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelSummary'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ModelSummary:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
          format: int64
        digest:
          type: string
        modified_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [available, downloading, error]
...
    GenerateRequest:
      type: object
      required:
        - model
        - prompt
      properties:
        model:
          type: string
        prompt:
          type: string
        max_tokens:
          type: integer
          minimum: 1
          maximum: 4096
          default: 100
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.8
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 0.9
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        stream:
          type: boolean
          default: false
    GenerateResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          enum: [text_completion]
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
        usage:
          $ref: '#/components/schemas/Usage'
    Choice:
      type: object
      properties:
        index:
          type: integer
        text:
          type: string
        finish_reason:
          type: string
          enum: [stop, length, temperature]
...
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...