# ollama-distributed/pkg/api/endpoints.go
package api

import (
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
)

// Health and Version endpoints

// health returns the health status of the server
func (s *Server) health(c *gin.Context) {
	status := "healthy"
	if !s.IsHealthy() {
		status = "unhealthy"
		c.JSON(http.StatusServiceUnavailable, gin.H{
			"status":    status,
			"timestamp": time.Now(),
		})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"status":    status,
		"timestamp": time.Now(),
		"uptime":    time.Since(time.Now().Add(-24 * time.Hour)), // Mock uptime
	})
}

// version returns the server version information
func (s *Server) version(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"version":     "1.0.0",
		"build_time":  "2024-01-15T10:00:00Z",
		"git_commit":  "abc123def456",
		"go_version":  "go1.21.0",
		"platform":    "linux/amd64",
		"api_version": "v1",
	})
}

// Model Management endpoints

// getModels returns all available models
func (s *Server) getModels(c *gin.Context) {
	s.LogRequest(c, "get_models")

	models := []map[string]interface{}{
		{
			"name":         "llama2:7b",
			"size":         3826793677,
			"digest":       "sha256:fe938a131f40e6f6d40083c9f0f430a515233eb2edaa6d72eb85c50d64f2300e",
			"modified_at":  time.Now().Add(-24 * time.Hour),
			"details": map[string]interface{}{
				"parent_model":    "",
				"format":         "gguf",
				"family":         "llama",
				"families":       []string{"llama"},
				"parameter_size": "7B",
				"quantization_level": "Q4_0",
			},
		},
		{
			"name":         "codellama:13b",
			"size":         7365960935,
			"digest":       "sha256:9f438cb9cd581fc025612d27f7c1a6669ff83a8bb0ed86c94fcf4c5440555697",
			"modified_at":  time.Now().Add(-48 * time.Hour),
			"details": map[string]interface{}{
				"parent_model":    "",
				"format":         "gguf",
				"family":         "llama",
				"families":       []string{"llama"},
				"parameter_size": "13B",
				"quantization_level": "Q4_0",
			},
		},
	}

	c.JSON(http.StatusOK, gin.H{
		"models": models,
	})
}

// getModel returns information about a specific model
func (s *Server) getModel(c *gin.Context) {
	modelName := c.Param("name")
	s.LogRequest(c, "get_model")

	// Mock model information
	model := map[string]interface{}{
		"name":         modelName,
		"size":         3826793677,
		"digest":       "sha256:fe938a131f40e6f6d40083c9f0f430a515233eb2edaa6d72eb85c50d64f2300e",
		"modified_at":  time.Now().Add(-24 * time.Hour),
		"details": map[string]interface{}{
			"parent_model":       "",
			"format":            "gguf",
			"family":            "llama",
			"families":          []string{"llama"},
			"parameter_size":    "7B",
			"quantization_level": "Q4_0",
		},
	}

	c.JSON(http.StatusOK, model)
}

// downloadModel downloads a model
func (s *Server) downloadModel(c *gin.Context) {
	modelName := c.Param("name")
	s.LogRequest(c, "download_model")

	// Mock download process
	c.JSON(http.StatusAccepted, gin.H{
		"status":  "downloading",
		"model":   modelName,
		"message": "Model download started",
	})
}

// deleteModel deletes a model
func (s *Server) deleteModel(c *gin.Context) {
	modelName := c.Param("name")
	s.LogRequest(c, "delete_model")

	// Mock deletion process
	c.JSON(http.StatusOK, gin.H{
		"status":  "deleted",
		"model":   modelName,
		"message": "Model deleted successfully",
	})
}

// Node Management endpoints

// getNodes returns all cluster nodes
func (s *Server) getNodes(c *gin.Context) {
	s.LogRequest(c, "get_nodes")

	nodes := []map[string]interface{}{
		{
			"id":           "node-001",
			"address":      "192.168.1.100:8080",
			"status":       "online",
			"last_seen":    time.Now().Add(-5 * time.Minute),
			"models":       []string{"llama2:7b", "codellama:13b"},
			"capacity": map[string]interface{}{
				"cpu_cores":    8,
				"memory_gb":    32,
				"disk_gb":      500,
				"gpu_count":    1,
			},
			"usage": map[string]interface{}{
				"cpu_usage":    0.45,
				"memory_usage": 0.67,
				"disk_usage":   0.23,
				"gpu_usage":    0.12,
			},
		},
		{
			"id":           "node-002",
			"address":      "192.168.1.101:8080",
			"status":       "online",
			"last_seen":    time.Now().Add(-2 * time.Minute),
			"models":       []string{"llama2:7b"},
			"capacity": map[string]interface{}{
				"cpu_cores":    16,
				"memory_gb":    64,
				"disk_gb":      1000,
				"gpu_count":    2,
			},
			"usage": map[string]interface{}{
				"cpu_usage":    0.23,
				"memory_usage": 0.34,
				"disk_usage":   0.15,
				"gpu_usage":    0.78,
			},
		},
	}

	c.JSON(http.StatusOK, gin.H{
		"nodes": nodes,
	})
}

// getNode returns information about a specific node
func (s *Server) getNode(c *gin.Context) {
	nodeID := c.Param("id")
	s.LogRequest(c, "get_node")

	// Mock node information
	node := map[string]interface{}{
		"id":           nodeID,
		"address":      "192.168.1.100:8080",
		"status":       "online",
		"last_seen":    time.Now().Add(-5 * time.Minute),
		"models":       []string{"llama2:7b", "codellama:13b"},
		"capacity": map[string]interface{}{
			"cpu_cores":    8,
			"memory_gb":    32,
			"disk_gb":      500,
			"gpu_count":    1,
		},
		"usage": map[string]interface{}{
			"cpu_usage":    0.45,
			"memory_usage": 0.67,
			"disk_usage":   0.23,
			"gpu_usage":    0.12,
		},
	}

	c.JSON(http.StatusOK, node)
}

// drainNode drains a node for maintenance
func (s *Server) drainNode(c *gin.Context) {
	nodeID := c.Param("id")
	s.LogRequest(c, "drain_node")

	c.JSON(http.StatusOK, gin.H{
		"status":  "draining",
		"node_id": nodeID,
		"message": "Node drain initiated",
	})
}

// undrainNode undoes node draining
func (s *Server) undrainNode(c *gin.Context) {
	nodeID := c.Param("id")
	s.LogRequest(c, "undrain_node")

	c.JSON(http.StatusOK, gin.H{
		"status":  "active",
		"node_id": nodeID,
		"message": "Node undrain completed",
	})
}

// Inference endpoints

// generate handles text generation requests
func (s *Server) generate(c *gin.Context) {
	s.LogRequest(c, "generate")

	var req GenerateRequest
	if err := s.ValidateRequest(c, &req); err != nil {
		return
	}

	// Mock generation response
	response := GenerateResponse{
		Model:     req.Model,
		Response:  "This is a mock response to your prompt: " + req.Prompt,
		Done:      true,
		Context:   []int{1, 2, 3, 4, 5},
		CreatedAt: time.Now(),
		TotalDuration: 1500000000, // 1.5 seconds in nanoseconds
		LoadDuration:  500000000,  // 0.5 seconds
		PromptEvalCount: 25,
		PromptEvalDuration: 200000000, // 0.2 seconds
		EvalCount: 50,
		EvalDuration: 800000000, // 0.8 seconds
	}

	c.JSON(http.StatusOK, response)
}

// chat handles chat completion requests
func (s *Server) chat(c *gin.Context) {
	s.LogRequest(c, "chat")

	var req ChatRequest
	if err := s.ValidateRequest(c, &req); err != nil {
		return
	}

	// Mock chat response
	response := ChatResponse{
		Model:     req.Model,
		Message: Message{
			Role:    "assistant",
			Content: "This is a mock chat response.",
		},
		Done:      true,
		CreatedAt: time.Now(),
		TotalDuration: 1200000000, // 1.2 seconds
		LoadDuration:  300000000,  // 0.3 seconds
		PromptEvalCount: 20,
		PromptEvalDuration: 150000000, // 0.15 seconds
		EvalCount: 40,
		EvalDuration: 750000000, // 0.75 seconds
	}

	c.JSON(http.StatusOK, response)
}

// embeddings handles embedding generation requests
func (s *Server) embeddings(c *gin.Context) {
	s.LogRequest(c, "embeddings")

	var req EmbeddingRequest
	if err := s.ValidateRequest(c, &req); err != nil {
		return
	}

	// Mock embedding response
	embedding := make([]float64, 4096) // Mock 4096-dimensional embedding
	for i := range embedding {
		embedding[i] = float64(i) * 0.001
	}

	response := EmbeddingResponse{
		Model:     req.Model,
		Embedding: embedding,
	}

	c.JSON(http.StatusOK, response)
}

// Cluster Management endpoints

// getClusterStatus returns cluster status
func (s *Server) getClusterStatus(c *gin.Context) {
	s.LogRequest(c, "get_cluster_status")

	status := map[string]interface{}{
		"cluster_id":    "ollama-cluster-001",
		"leader_node":   "node-001",
		"total_nodes":   2,
		"healthy_nodes": 2,
		"total_models":  2,
		"status":        "healthy",
		"last_updated":  time.Now(),
	}

	c.JSON(http.StatusOK, status)
}

// getClusterLeader returns cluster leader information
func (s *Server) getClusterLeader(c *gin.Context) {
	s.LogRequest(c, "get_cluster_leader")

	leader := map[string]interface{}{
		"node_id":   "node-001",
		"address":   "192.168.1.100:8080",
		"term":      5,
		"since":     time.Now().Add(-2 * time.Hour),
	}

	c.JSON(http.StatusOK, leader)
}

// joinCluster handles cluster join requests
func (s *Server) joinCluster(c *gin.Context) {
	s.LogRequest(c, "join_cluster")

	var req JoinClusterRequest
	if err := s.ValidateRequest(c, &req); err != nil {
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"status":  "joined",
		"message": "Successfully joined cluster",
		"node_id": req.NodeID,
	})
}

// leaveCluster handles cluster leave requests
func (s *Server) leaveCluster(c *gin.Context) {
	s.LogRequest(c, "leave_cluster")

	c.JSON(http.StatusOK, gin.H{
		"status":  "left",
		"message": "Successfully left cluster",
	})
}

// Transfer Management endpoints

// getTransfers returns all model transfers
func (s *Server) getTransfers(c *gin.Context) {
	s.LogRequest(c, "get_transfers")

	transfers := []map[string]interface{}{
		{
			"id":          "transfer-001",
			"model":       "llama2:7b",
			"source":      "node-001",
			"destination": "node-002",
			"status":      "completed",
			"progress":    100,
			"started_at":  time.Now().Add(-1 * time.Hour),
			"completed_at": time.Now().Add(-30 * time.Minute),
		},
		{
			"id":          "transfer-002",
			"model":       "codellama:13b",
			"source":      "node-001",
			"destination": "node-003",
			"status":      "in_progress",
			"progress":    45,
			"started_at":  time.Now().Add(-15 * time.Minute),
		},
	}

	c.JSON(http.StatusOK, gin.H{
		"transfers": transfers,
	})
}

// getTransfer returns information about a specific transfer
func (s *Server) getTransfer(c *gin.Context) {
	transferID := c.Param("id")
	s.LogRequest(c, "get_transfer")

	transfer := map[string]interface{}{
		"id":          transferID,
		"model":       "llama2:7b",
		"source":      "node-001",
		"destination": "node-002",
		"status":      "completed",
		"progress":    100,
		"started_at":  time.Now().Add(-1 * time.Hour),
		"completed_at": time.Now().Add(-30 * time.Minute),
	}

	c.JSON(http.StatusOK, transfer)
}

// cancelTransfer cancels a model transfer
func (s *Server) cancelTransfer(c *gin.Context) {
	transferID := c.Param("id")
	s.LogRequest(c, "cancel_transfer")

	c.JSON(http.StatusOK, gin.H{
		"status":      "cancelled",
		"transfer_id": transferID,
		"message":     "Transfer cancelled successfully",
	})
}

// Distribution Management endpoints

// autoConfigureDistribution automatically configures model distribution
func (s *Server) autoConfigureDistribution(c *gin.Context) {
	s.LogRequest(c, "auto_configure_distribution")

	c.JSON(http.StatusOK, gin.H{
		"status":  "configured",
		"message": "Auto-distribution configured successfully",
		"policies": []map[string]interface{}{
			{
				"model":        "llama2:7b",
				"min_replicas": 2,
				"max_replicas": 4,
				"strategy":     "balanced",
			},
			{
				"model":        "codellama:13b",
				"min_replicas": 1,
				"max_replicas": 2,
				"strategy":     "performance",
			},
		},
	})
}

// System endpoints

// getMetrics returns system metrics
func (s *Server) getMetrics(c *gin.Context) {
	// Return Prometheus-style metrics
	metrics := `# HELP ollama_requests_total Total number of requests
# TYPE ollama_requests_total counter
ollama_requests_total{method="GET",endpoint="/api/v1/models"} 150
ollama_requests_total{method="POST",endpoint="/api/v1/generate"} 1240

# HELP ollama_request_duration_seconds Request duration in seconds
# TYPE ollama_request_duration_seconds histogram
ollama_request_duration_seconds_bucket{endpoint="/api/v1/generate",le="0.1"} 245
ollama_request_duration_seconds_bucket{endpoint="/api/v1/generate",le="0.5"} 890
ollama_request_duration_seconds_bucket{endpoint="/api/v1/generate",le="1.0"} 1180
ollama_request_duration_seconds_bucket{endpoint="/api/v1/generate",le="+Inf"} 1240

# HELP ollama_nodes_total Total number of cluster nodes
# TYPE ollama_nodes_total gauge
ollama_nodes_total 2

# HELP ollama_models_total Total number of models
# TYPE ollama_models_total gauge
ollama_models_total 2
`

	c.Header("Content-Type", "text/plain; version=0.0.4; charset=utf-8")
	c.String(http.StatusOK, metrics)
}

// getStats returns server statistics
func (s *Server) getStats(c *gin.Context) {
	s.LogRequest(c, "get_stats")

	stats := s.GetStats()
	c.JSON(http.StatusOK, stats)
}

// getConfig returns server configuration
func (s *Server) getConfig(c *gin.Context) {
	s.LogRequest(c, "get_config")

	config := map[string]interface{}{
		"api": map[string]interface{}{
			"listen":     "0.0.0.0:8080",
			"tls_enabled": false,
		},
		"cluster": map[string]interface{}{
			"id":       "ollama-cluster-001",
			"max_nodes": 10,
		},
		"models": map[string]interface{}{
			"auto_pull":     true,
			"max_models":    50,
			"cache_size_gb": 100,
		},
	}

	c.JSON(http.StatusOK, config)
}

// updateConfig updates server configuration
func (s *Server) updateConfig(c *gin.Context) {
	s.LogRequest(c, "update_config")

	var req map[string]interface{}
	if err := s.ValidateRequest(c, &req); err != nil {
		return
	}

	// Mock configuration update
	c.JSON(http.StatusOK, gin.H{
		"status":  "updated",
		"message": "Configuration updated successfully",
	})
}

// Dashboard endpoints

// getDashboardData returns dashboard data
func (s *Server) getDashboardData(c *gin.Context) {
	s.LogRequest(c, "get_dashboard_data")

	data := map[string]interface{}{
		"cluster_status": "healthy",
		"total_nodes":    2,
		"total_models":   2,
		"active_requests": 5,
		"total_requests": 1240,
		"uptime":        "24h 15m",
		"recent_activity": []map[string]interface{}{
			{
				"timestamp": time.Now().Add(-5 * time.Minute),
				"event":     "Model generation completed",
				"user":      "user@example.com",
				"model":     "llama2:7b",
			},
			{
				"timestamp": time.Now().Add(-10 * time.Minute),
				"event":     "Node joined cluster",
				"node":      "node-003",
			},
		},
	}

	c.JSON(http.StatusOK, data)
}

// getDashboardMetrics returns dashboard metrics
func (s *Server) getDashboardMetrics(c *gin.Context) {
	s.LogRequest(c, "get_dashboard_metrics")

	metrics := map[string]interface{}{
		"requests_per_minute": 15.7,
		"average_latency":     1.2,
		"error_rate":         0.02,
		"cpu_usage":          45.6,
		"memory_usage":       67.8,
		"disk_usage":         23.4,
		"network_io": map[string]interface{}{
			"inbound_mbps":  12.5,
			"outbound_mbps": 8.3,
		},
	}

	c.JSON(http.StatusOK, metrics)
}

// Notification endpoints

// getNotifications returns user notifications
func (s *Server) getNotifications(c *gin.Context) {
	userID := c.GetString("user_id")
	s.LogRequest(c, "get_notifications")

	notifications := []map[string]interface{}{
		{
			"id":         "notif-001",
			"type":       "info",
			"title":      "Model Download Complete",
			"message":    "Model 'llama2:7b' has been downloaded successfully",
			"read":       false,
			"created_at": time.Now().Add(-30 * time.Minute),
		},
		{
			"id":         "notif-002",
			"type":       "warning",
			"title":      "High Memory Usage",
			"message":    "Node 'node-001' is experiencing high memory usage (85%)",
			"read":       true,
			"created_at": time.Now().Add(-2 * time.Hour),
		},
	}

	c.JSON(http.StatusOK, gin.H{
		"notifications": notifications,
		"user_id":       userID,
	})
}

// markNotificationRead marks a notification as read
func (s *Server) markNotificationRead(c *gin.Context) {
	notificationID := c.Param("id")
	s.LogRequest(c, "mark_notification_read")

	c.JSON(http.StatusOK, gin.H{
		"status":          "updated",
		"notification_id": notificationID,
		"message":         "Notification marked as read",
	})
}

// Request/Response types for API endpoints

type GenerateRequest struct {
	Model    string `json:"model" binding:"required"`
	Prompt   string `json:"prompt" binding:"required"`
	Stream   bool   `json:"stream,omitempty"`
	Options  map[string]interface{} `json:"options,omitempty"`
	Context  []int  `json:"context,omitempty"`
}

type GenerateResponse struct {
	Model     string    `json:"model"`
	Response  string    `json:"response"`
	Done      bool      `json:"done"`
	Context   []int     `json:"context,omitempty"`
	CreatedAt time.Time `json:"created_at"`
	TotalDuration     int64 `json:"total_duration"`
	LoadDuration      int64 `json:"load_duration"`
	PromptEvalCount   int   `json:"prompt_eval_count"`
	PromptEvalDuration int64 `json:"prompt_eval_duration"`
	EvalCount         int   `json:"eval_count"`
	EvalDuration      int64 `json:"eval_duration"`
}

type Message struct {
	Role    string `json:"role"`
	Content string `json:"content"`
}

type ChatRequest struct {
	Model    string    `json:"model" binding:"required"`
	Messages []Message `json:"messages" binding:"required"`
	Stream   bool      `json:"stream,omitempty"`
	Options  map[string]interface{} `json:"options,omitempty"`
}

type ChatResponse struct {
	Model     string    `json:"model"`
	Message   Message   `json:"message"`
	Done      bool      `json:"done"`
	CreatedAt time.Time `json:"created_at"`
	TotalDuration     int64 `json:"total_duration"`
	LoadDuration      int64 `json:"load_duration"`
	PromptEvalCount   int   `json:"prompt_eval_count"`
	PromptEvalDuration int64 `json:"prompt_eval_duration"`
	EvalCount         int   `json:"eval_count"`
	EvalDuration      int64 `json:"eval_duration"`
}

type EmbeddingRequest struct {
	Model  string `json:"model" binding:"required"`
	Prompt string `json:"prompt" binding:"required"`
}

type EmbeddingResponse struct {
	Model     string    `json:"model"`
	Embedding []float64 `json:"embedding"`
}

type JoinClusterRequest struct {
	NodeID  string `json:"node_id" binding:"required"`
	Address string `json:"address" binding:"required"`
}